Restructured entire project to use a standard folder-index naming convention. Switched all of the scss over to css-modules because was beginning to run into namespace clashes
