+++ gatsby-browser.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from './src/components/layout/layout';
+import Layout from './src/components/layout';
 import './src/styles/global.scss';
 
 export function wrapPageElement({ element, props }) {
