+++ src/components/footer/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import './footer.scss'
+import { footerComponent } from './footer.module.scss'
 
 export default function Footer(props: any): React.ReactComponentElement<any> {
 
     return (
-        <div className="footer-component">
+        <div className={footerComponent}>
             <p className='center'>&copy; Michael E Stallings {new Date().getFullYear()}</p>
         </div>
     )
