+++ src/components/logo/logo.module.scss
@@ -11,7 +11,7 @@
         display: block;
         text-shadow: 0.18em 0.18em 0 rgba(0, 0, 0, 0.05);
         perspective: 100px;
-        letter-spacing: 0.08em;
+        letter-spacing: 0.1em;
 
         .letter {
             // Variable declarations
@@ -22,7 +22,7 @@
             --rotateX: 0deg;
             //////////////////
             // font-size: 3rem;
-            font-size: 3.2rem;
+            font-size: 3.5rem;
             line-height: 1;
             display: inline-block;      // Without this, no transforms will work.
             transform: scale(var(--scale)) translateY(var(--translateY));
